split BucketPreferences to clean data functions and react context provider
