+++ catalog/app/utils/BucketPreferences/BucketPreferences.ts
@@ -1,18 +1,12 @@
 import * as R from 'ramda'
-import * as React from 'react'
 
 import bucketPreferencesSchema from 'schemas/bucketConfig.yml.json'
 
-import * as AWS from 'utils/AWS'
-import * as Config from 'utils/Config'
-import { useData } from 'utils/Data'
-import * as Sentry from 'utils/Sentry'
+import * as bucketErrors from 'containers/Bucket/errors'
 import { makeSchemaValidator } from 'utils/json-schema'
 import yaml from 'utils/yaml'
-import * as bucketErrors from 'containers/Bucket/errors'
-import * as requests from 'containers/Bucket/requests'
 
-type SentryInstance = (command: 'captureMessage', message: string) => void
+export type SentryInstance = (command: 'captureMessage', message: string) => void
 
 export type ActionPreferences = Record<
   'copyPackage' | 'createPackage' | 'deleteRevision' | 'revisePackage',
@@ -47,11 +41,11 @@ interface BucketPreferencesYaml {
   ui?: UiPreferencesYaml
 }
 
-interface BucketPreferences {
+export interface BucketPreferences {
   ui: UiPreferences
 }
 
-const defaultPreferences: BucketPreferences = {
+export const defaultPreferences: BucketPreferences = {
   ui: {
     actions: {
       copyPackage: true,
@@ -77,28 +71,9 @@ const defaultPreferences: BucketPreferences = {
   },
 }
 
-const localModePreferences: BucketPreferences = {
-  ui: {
-    ...defaultPreferences.ui,
-    actions: {
-      copyPackage: false,
-      createPackage: false,
-      deleteRevision: false,
-      revisePackage: false,
-    },
-    blocks: {
-      analytics: true,
-      browser: true,
-      code: true,
-      meta: true,
-    },
-    nav: {
-      files: true,
-      packages: true,
-      queries: false,
-    },
-  },
-}
+const S3_PREFIX = 's3://'
+const normalizeBucketName = (input: string) =>
+  input.startsWith(S3_PREFIX) ? input.slice(S3_PREFIX.length) : input
 
 const bucketPreferencesValidator = makeSchemaValidator(bucketPreferencesSchema)
 
@@ -106,10 +81,6 @@ function validate(data: unknown): asserts data is BucketPreferencesYaml {
   const errors = bucketPreferencesValidator(data)
   if (errors.length) throw new bucketErrors.BucketPreferencesInvalid({ errors })
 }
-
-const S3_PREFIX = 's3://'
-const normalizeBucketName = (input: string) =>
-  input.startsWith(S3_PREFIX) ? input.slice(S3_PREFIX.length) : input
 
 function parseSourceBuckets(
   sentry: SentryInstance,
@@ -133,7 +104,10 @@ function parseSourceBuckets(
   }
 }
 
-function parse(bucketPreferencesYaml: string, sentry: SentryInstance): BucketPreferences {
+export function parse(
+  bucketPreferencesYaml: string,
+  sentry: SentryInstance,
+): BucketPreferences {
   const data = yaml(bucketPreferencesYaml)
   if (!data) return defaultPreferences
 
@@ -148,68 +122,3 @@ function parse(bucketPreferencesYaml: string, sentry: SentryInstance): BucketPre
     },
   }
 }
-
-const BUCKET_PREFERENCES_PATH = [
-  '.quilt/catalog/config.yaml',
-  '.quilt/catalog/config.yml',
-]
-
-interface FetchBucketPreferencesArgs {
-  s3: $TSFixMe
-  bucket: string
-  sentry: SentryInstance
-  local: boolean
-}
-
-async function fetchBucketPreferences({
-  s3,
-  sentry,
-  bucket,
-  local,
-}: FetchBucketPreferencesArgs) {
-  if (local) return localModePreferences
-  try {
-    const response = await requests.fetchFile({
-      s3,
-      bucket,
-      path: BUCKET_PREFERENCES_PATH,
-    })
-    return parse(response.Body.toString('utf-8'), sentry)
-  } catch (e) {
-    if (
-      e instanceof bucketErrors.FileNotFound ||
-      e instanceof bucketErrors.VersionNotFound
-    )
-      return defaultPreferences
-
-    // eslint-disable-next-line no-console
-    console.log('Unable to fetch')
-    // eslint-disable-next-line no-console
-    console.error(e)
-    throw e
-  }
-}
-
-const Ctx = React.createContext<BucketPreferences | null>(null)
-
-type ProviderProps = React.PropsWithChildren<{ bucket: string }>
-
-export function Provider({ bucket, children }: ProviderProps) {
-  const cfg = Config.use()
-  const local = cfg.mode === 'LOCAL'
-  const sentry = Sentry.use()
-  const s3 = AWS.S3.use()
-  const data = useData(fetchBucketPreferences, { s3, sentry, bucket, local })
-
-  // XXX: consider returning AsyncResult or using Suspense
-  const preferences = data.case({
-    Ok: R.identity,
-    Err: () => defaultPreferences,
-    _: () => null,
-  })
-  return <Ctx.Provider value={preferences}>{children}</Ctx.Provider>
-}
-
-export const useBucketPreferences = () => React.useContext(Ctx)
-
-export const use = useBucketPreferences
